pyproject.toml
[project]
name = "marketplace-api"
version = "0.1.0"
description = ""
authors = [
    {name = "x1ryrg",email = "zhdamarov-vladimir@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "ipython (>=9.0.2,<10.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "celery (>=5.5.0,<6.0.0)",
    "eventlet (>=0.39.1,<0.40.0)",
    "redis (>=5.2.1,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "yookassa (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
